requirements.txt
# Core dependencies
yt-dlp>=2023.12.30
faster-whisper>=0.10.0
openai>=1.3.0

# Audio and video processing
ffmpeg-python>=0.2.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
click>=8.1.0
rich>=13.0.0

# PDF generation
reportlab>=4.0.0
markdown>=3.4.0

# Development dependencies (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Google Drive API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
